# faster dense kron