Simple optimization with high reward for dense `kron`. Saves one line ;-)


MASTER:

```
a=rand(20,20);

julia> @benchmark kron($a,$a)
BenchmarkTools.Trial: 
  memory estimate:  1.22 MiB
  allocs estimate:  2
  --------------
  minimum time:     124.824 μs (0.00% GC)
  median time:      134.279 μs (0.00% GC)
  mean time:        207.965 μs (8.57% GC)
  maximum time:     27.515 ms (98.35% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

PR

```
julia> @benchmark kron($a,$a)
BenchmarkTools.Trial: 
  memory estimate:  1.22 MiB
  allocs estimate:  2
  --------------
  minimum time:     64.399 μs (0.00% GC)
  median time:      84.397 μs (0.00% GC)
  mean time:        112.454 μs (11.47% GC)
  maximum time:     28.646 ms (98.90% GC)
  --------------
  samples:          10000
  evals/sample:     1
```