@abraunst 2019-01-19
> How much of this is due to the inbounds and how much due to changing `m`?

About 15% due to the `m` change, the rest is all `@inbounds`